30a7d380ec409c3e17e6e7ebd2f02950cfcafb13 [bitreq] Support pipelining requests (Matt Corallo)
eaf64207367e16f6daf663c9f9a7106f6cacbd2a [bitreq] Add a `Client` which allows for connection reuse (Matt Corallo)
c26416cb5764a8039eafb2018dc6d95db51ce1fc [bitreq] Reuse the existing `AsyncConnection` on redirection (Matt Corallo)
704270cdae77600213656dc2f8dae653d9fb27e0 [bitreq] Reduce async test latency marginally (Matt Corallo)
996f4ffa9acf2516a61e3a30ef2bcb96479586b9 [bitreq] Include HTTP `Connection` headers on some test responses (Matt Corallo)
6049f36536fd06b0845da7380c9ae8074188bb6e [bitreq] Add an owned version of the `ConnectionParams` (Matt Corallo)
33f162c12ed1309cf2da53dd7a416184224ae2c2 [bitreq] Connect in `Connection`'s constructor not in `send` (Matt Corallo)
7f6fd458d130daee73ff87df2654240a2105f60c [bitreq] Remove kinda-reachable `unwrap` (Matt Corallo)
eaf6b5cd119995e3bb17ba58c28c8f5917eb2df0 [bitreq] Connect in `AsyncConnection`'s constructor not in `send` (Matt Corallo)
92d1b5be3c7aef514ae93559886f0aad7ec3c5c7 [bitreq] Correct port bit length (Matt Corallo)

Pull request description:

  I went into this thinking it would be easy and claude would be able to do it. I was very, very wrong. The `Connection` stuff was an absolute mess and basically had to be rewritten, but then actually properly supporting pipelining requests was more annoying that I thought. Still, it works now, and I even tested it in BDK's rust-esplora-client inside of ldk-node, which hasn't yet broken. Probably best to review `connection.rs` from scratch as of the third-to-last commit and as of the end and not bother looking at the earlier intermediate commits in `connection.rs`.

ACKs for top commit:
  tcharding:
    ACK 30a7d380ec409c3e17e6e7ebd2f02950cfcafb13

Tree-SHA512: 9d131bfe1346098366a0f91794b1225051ab76b32c2a832eb1b42605c82e9b72269f060318f3b0f26adc36226b6e5423e4d078ec9faba4b64a36bbcfe6baad2d

+++ bitreq/src/http_url.rs
@@ -5,16 +5,16 @@ use core::fmt::{self, Write};
 use crate::Error;
 
 #[cfg(feature = "std")]
-#[derive(Clone, Copy, PartialEq)]
+#[derive(Clone, Copy, Debug, Hash, PartialEq, Eq)]
 pub(crate) enum Port {
     ImplicitHttp,
     ImplicitHttps,
-    Explicit(u32),
+    Explicit(u16),
 }
 
 #[cfg(feature = "std")]
 impl Port {
-    pub(crate) fn port(self) -> u32 {
+    pub(crate) fn port(self) -> u16 {
         match self {
             Port::ImplicitHttp => 80,
             Port::ImplicitHttps => 443,
@@ -145,7 +145,7 @@ impl HttpUrl {
         }
 
         // Set appropriate port
-        let port = port.parse::<u32>().map(Port::Explicit).unwrap_or_else(|_| {
+        let port = port.parse::<u16>().map(Port::Explicit).unwrap_or_else(|_| {
             if https {
                 Port::ImplicitHttps
             } else {
